Update documentation (#713)
## What do these changes do?

This PR updates the documentation to reflect the latest changes:
recommends to use the new `aclose()` method (ad2b1bd) and clarifies that
the slowdown "can be", but not necessarily "will be" (4a57895). In fact,
in no-wait tests Janus can now be comparable to standard queues, but in
worst cases where notification methods are often called, it can still
show significant performance slowdown, so "only when used as intended"
remains. In particular, in Culsans benchmarks Janus is still slower than
the naive solution.

## Are there changes in behavior for the user?

There are no behavior changes for users.